pyproject.toml
[tool.poetry]
name = "upskill-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
SQLAlchemy = "^1.4.34"
alembic = "^1.7.7"
fastapi = "^0.75.1"
uvicorn = "^0.17.6"
python-dotenv = "^0.20.0"
psycopg2-binary = "^2.9.3"
databases = "^0.5.5"
asyncpg = "^0.25.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
